added coredata logic

+++ KC_PhotoArts+/CartData.swift
@@ -15,19 +15,19 @@ import UIKit
 class CartData: NSObject {
 
     static var sharedInstance = [CartData]() /*https://stackoverflow.com/questions/27651507/passing-data-between-tab-viewed-controllers-in-swift */ //to store cart items
-    static var totalPrice:Int = 0 //to store total price of items in the cart
+    static var totalPrice:Int64 = 0 //to store total price of items in the cart
     
     
-    var quantity:Int!
+    var quantity:Int64!
     var size:String!
     var frame:String!
-    var itemPrice:Int!
-    var itemTotal:Int!
+    var itemPrice:Int64!
+    var itemTotal:Int64!
     var itemNumber:String!
     var itemName:String!
     var itemImageURL:String!
     
-    init(quantity:Int, size:String, frame:String, itemPrice:Int, itemTotal:Int, itemNumber:String, itemName:String, itemImageURL:String)
+    init(quantity:Int64, size:String, frame:String, itemPrice:Int64, itemTotal:Int64, itemNumber:String, itemName:String, itemImageURL:String)
     {
         self.quantity = quantity
         self.frame = frame
